# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "GabomaPay",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://oask.ca/api"
        }
    ],
    "paths": {
...
        "/admin/users/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION_UTILISATEUR"
                ],
                "summary": "Renvoie les informations d'un utilisateur par son ID",
                "operationId": "085e470d12a568ddd14f5a29270456ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/users/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION_UTILISATEUR"
                ],
                "summary": "Active ou désactive l'état d'un utilisateur par son ID",
                "operationId": "f43481b9e665f9e8f752b7c8b8dd8c06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}